move events => event-series
