+++ common/services/prismic/event-series.js
