+++ common/services/prismic/parsers.js
@@ -12,7 +12,7 @@ import type { CaptionedImage } from '../../model/captioned-image';
 import type { ImagePromo } from '../../model/image-promo';
 import { licenseTypeArray } from '../../model/license';
 import { parsePage } from './pages';
-import { parseEventSeries } from './events';
+import { parseEventSeries } from './event-series';
 
 const placeHolderImage = {
   contentUrl: 'https://via.placeholder.com/1600x900?text=%20',
